Introduce file record generation; change size field to BIGINT; go down to 5000 objects

+++ src/main/resources/schema.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS digital_object_file (
     identifier VARCHAR(255) NOT NULL,
     file_format VARCHAR(255) NOT NULL,
     file_function VARCHAR(255) NOT NULL,
-    "size" INTEGER NOT NULL,
+    "size" BIGINT NOT NULL,
     digest VARCHAR(255) NOT NULL,
     created_at TIMESTAMP NOT NULL,
     digital_object_version BIGSERIAL REFERENCES digital_object_version (id)
@@ -49,7 +49,7 @@ CREATE TABLE IF NOT EXISTS object_file (
     identifier VARCHAR(255) NOT NULL,
     file_format VARCHAR(255) NOT NULL,
     file_function VARCHAR(255) NOT NULL,
-    "size" INTEGER NOT NULL,
+    "size" BIGINT NOT NULL,
     digest VARCHAR(255) NOT NULL,
     created_at TIMESTAMP NOT NULL,
     intellectual_object BIGSERIAL REFERENCES intellectual_object (id)
